Rename file and some tweaks
