+++ src/Storages/MergeTree/BackgroundProcessList.h
@@ -9,6 +9,8 @@
 namespace DB
 {
 
+/// Common code for background processes lists, like system.merges and system.fetches
+/// Look at examples in MergeList and ReplicatedFetchList
 
 template <typename ListElement, typename Info>
 class BackgroundProcessList;
