+++ src/Storages/System/StorageSystemFetches.h
@@ -10,7 +10,7 @@ namespace DB
 
 class Context;
 
-/// system.fetches table. Takes data from context.getReplicatedFetchesList()
+/// system.fetches table. Takes data from context.getReplicatedFetchList()
 class StorageSystemFetches final : public ext::shared_ptr_helper<StorageSystemFetches>, public IStorageSystemOneBlock<StorageSystemFetches >
 {
     friend struct ext::shared_ptr_helper<StorageSystemFetches>;
